refactor: main 과 test 내 패키지 변경 및 클래스 이동
