+++ src/main/java/com/kakao/cafe/domain/User.java
@@ -1,4 +1,4 @@
-package com.kakao.cafe;
+package com.kakao.cafe.domain;
 
 public class User {
