+++ src/main/java/com/kakao/cafe/repository/UserRepository.java
@@ -1,0 +1,17 @@
+package com.kakao.cafe.repository;
+
+import com.kakao.cafe.domain.User;
+import java.util.List;
+import java.util.Optional;
+
+public interface UserRepository {
+
+    User save(User user);
+
+    List<User> findAll();
+
+    Optional<User> findByUserId(String userId);
+
+    void deleteAll();
+
+}
