+++ src/test/java/com/kakao/cafe/integration/UserSetUp.java
@@ -1,7 +1,7 @@
 package com.kakao.cafe.integration;
 
-import com.kakao.cafe.User;
-import com.kakao.cafe.UserRepository;
+import com.kakao.cafe.domain.User;
+import com.kakao.cafe.repository.UserCollectionRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -9,14 +9,13 @@ import org.springframework.stereotype.Component;
 public class UserSetUp {
 
     @Autowired
-    private UserRepository userRepository;
+    private UserCollectionRepository userRepository;
 
     public User saveUser(User user) {
         return userRepository.save(user);
     }
 
     public void rollback() {
-        userRepository.setUserNum(0);
         userRepository.deleteAll();
     }
 }
